chore: join create and edit into single form component
